example2/django/todolist/requirements.txt
Django==3.2.13
djangorestframework==3.13.1